programs/tts/silero_tts/requirements.txt
torchaudio
omegaconf
numpy